[Refact] 임시 수정
