+++ .github/workflows/gradle.yml
@@ -2,7 +2,7 @@ name: Auth Server CI/CD
 
 on:
   push:
-    branches: [ "main" ,"dev", "feature/refact"]  # main push 시 배포
+    branches: [ "main" ,"dev"]  # main push 시 배포
 
 jobs:
   build-and-deploy:
