+++ src/main/java/com/chat_server/user/enums/UserType.java
@@ -1,10 +1,8 @@
-package com.chat_server.user.domain;
-
-import lombok.Getter;
+package com.chat_server.user.enums;
 
 /**
  * packageName    : com.chat_server.user.domain
- * fileName       : UserSstatus
+ * fileName       : UserType
  * author         : parkminsu
  * date           : 25. 5. 26.
  * description    :
@@ -13,17 +11,19 @@ import lombok.Getter;
  * -----------------------------------------------------------
  * 25. 5. 26.        parkminsu       최초 생성
  */
-@Getter
-public enum UserStatus {
-    ACTIVE("활성 사용자"),
-    DORMANT("휴면 사용자"),
-    BLOCKED("차단된 사용자"),
-    WITHDRAWN("탈퇴한 사용자");
+
+public enum UserType {
+    USER("일반 사용자"),
+    ADMIN("관리자");
 
     private final String description;
 
-    UserStatus(String description) {
+    UserType(String description) {
         this.description = description;
     }
 
+    public String getDescription() {
+        return description;
+    }
 }
+
