+++ src/main/java/com/chat_server/user/service/impl/UserServiceImpl.java
@@ -4,7 +4,7 @@ import com.chat_server.security.dto.UserPrincipal;
 import com.chat_server.gender.entity.Gender;
 import com.chat_server.gender.exception.GenderNotFoundException;
 import com.chat_server.gender.repository.GenderRepository;
-import com.chat_server.user.domain.UserType;
+import com.chat_server.user.enums.UserType;
 import com.chat_server.user.dto.request.UserRegisterRequest;
 import com.chat_server.user.dto.response.UserAuthenticationResponse;
 import com.chat_server.user.entity.User;
@@ -14,9 +14,7 @@ import com.chat_server.user.repository.UserRepository;
 import com.chat_server.user.service.UserService;
 import com.chat_server.userprofile.enrtity.UserProfile;
 import com.chat_server.userprofile.repository.UserProfileRepository;
-import com.chat_server.userstatus.entity.UserStatus;
-import com.chat_server.userstatus.exception.UserStatusNotFoundException;
-import com.chat_server.userstatus.repository.UserStatusRepository;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
